+++ src/entities/plansCard/PlansCard.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/widgets/button'
+import { Button } from '@/shared/ui/button'
 import { PlansCardItem } from '@/entities/plansCard/config/types'
 
 import IconCheck from '@/shared/assets/check-bold.svg'
